requirements.txt
streamlit
Pillow
matplotlib
wordcloud
openpyxl